+++ src/main/java/com/jelly/zzirit/domain/order/service/OrderCancellationFacade.java
@@ -3,6 +3,7 @@ package com.jelly.zzirit.domain.order.service;
 import com.jelly.zzirit.domain.member.entity.Member;
 import com.jelly.zzirit.domain.order.entity.Order;
 import com.jelly.zzirit.domain.order.repository.OrderRepository;
+import com.jelly.zzirit.domain.order.service.pay.RefundService;
 import com.jelly.zzirit.global.authorization.AuthorizationService;
 import com.jelly.zzirit.global.exception.custom.InvalidOrderException;
 import lombok.RequiredArgsConstructor;
