+++ src/main/java/com/jelly/zzirit/domain/order/service/pay/RefundService.java
@@ -1,6 +1,9 @@
 package com.jelly.zzirit.domain.order.service.pay;
 
 import java.math.BigDecimal;
+import java.nio.charset.StandardCharsets;
+import java.util.Base64;
+import java.util.HashMap;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Value;
@@ -53,6 +56,37 @@ public class RefundService {
 		}
 	}
 
+	/**
+	 * 주문 취소 요청에 따른 결제 취소
+	 * @param orderId 취소할 주문의 아이디
+	 * @param paymentKey 결제 정보 키
+	 */
+	public void refund(Long orderId, String paymentKey) {
+		String url = "https://api.tosspayments.com/v1/payments/" + paymentKey + "/cancel";
+
+		HttpHeaders headers = getHeaders(); // 헤더 생성
+		Map<String, Object> body = getBody(); // 바디 생성
+
+		HttpEntity<Map<String, Object>> request = new HttpEntity<>(body, headers);
+
+		try {
+			ResponseEntity<String> response = restTemplate.exchange(
+				url,
+				HttpMethod.POST,
+				request,
+				String.class
+			);
+
+			if (!response.getStatusCode().is2xxSuccessful()) {
+				log.error("환불 실패: status={}, body={}", response.getStatusCode(), response.getBody());
+				throw new IllegalStateException("토스 결제 취소 API 실패");
+			}
+		} catch (Exception e) {
+			log.error("환불 실패: orderId={}, message={}", orderId, e.getMessage(), e);
+			throw new IllegalStateException("환불 처리 실패");
+		}
+	}
+
 	private void requestTossRefund(String paymentKey, BigDecimal amount) {
 		String url = "https://api.tosspayments.com/v1/payments/" + paymentKey + "/cancel";
 
@@ -72,4 +106,24 @@ public class RefundService {
 			throw new InvalidOrderException(BaseResponseStatus.TOSS_REFUND_FAILED);
 		}
 	}
+
+	private HttpHeaders getHeaders() {
+		HttpHeaders headers = new HttpHeaders();
+		String encodedKey = Base64.getEncoder().encodeToString((secretKey + ":").getBytes(StandardCharsets.UTF_8));
+
+		headers.setContentType(MediaType.APPLICATION_JSON);
+		headers.set("Authorization", "Basic " + encodedKey);
+
+		return headers;
+	}
+
+	private Map<String, Object> getBody() {
+		String cancelReason = "구매자 변심으로 인한 주문 취소";
+		Map<String, Object> body = new HashMap<>();
+
+		body.put("cancelReason", cancelReason);
+
+		return body;
+	}
+
 }
